# backend/docs/swagger.yaml
definitions:
  ApplicationCreation:
    properties:
      name:
        type: string
      phone_number:
        type: string
    type: object
  ApplicationRead:
    properties:
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  ErrorResponseModel:
    properties:
      error:
        type: string
    type: object
  LoginModel:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  RegistrationModel:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /application:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all applications from the system
      produces:
      - application/json
      responses:
        "200":
          description: List of all applications
          schema:
            items:
              $ref: '#/definitions/ApplicationRead'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get all applications
      tags:
      - application
    post:
      consumes:
      - application/json
      description: Add a new application to the system with the provided details
      parameters:
      - description: Application details
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/ApplicationCreation'
      produces:
      - application/json
      responses:
        "201":
          description: Application created successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Create a new application
      tags:
      - application
  /application/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an application from the system using its unique identifier
      parameters:
      - description: ID of the application to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Application deleted successfully
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Delete an application
      tags:
      - application
    get:
      consumes:
      - application/json
      description: Retrieve a specific application by its unique identifier
      parameters:
      - description: ID of the application to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Application details
          schema:
            $ref: '#/definitions/ApplicationRead'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get application by ID
      tags:
      - application
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token as a cookie
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/LoginModel'
      produces:
      - application/json
      responses:
        "200":
          description: User ID
          schema:
            type: string
        "401":
          description: Already logged in
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: User login
      tags:
      - auth
  /auth/registration:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/RegistrationModel'
      produces:
      - application/json
      responses:
        "200":
          description: User ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: User registration
      tags:
      - auth
...
swagger: "2.0"